% Auto-generated: do not edit by hand
\name{'govuk'Accordion}

\alias{'govuk'Accordion}

\title{Accordion component}

\description{
Lazy loaded ComponentTemplate  @param { 	id,  accordionContent, } [props={}] @return {*}
}

\usage{
'govuk'Accordion(id=NULL, accordionContent=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{accordionContent}{List of lists containing elements 'heading', 'children'.
those elements have the following types:
  - heading (character; optional)
  - children (unnamed list; optional)s. Accordion heading and children content object.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

