% Auto-generated: do not edit by hand
\name{'govuk'AutoComplete}

\alias{'govuk'AutoComplete}

\title{AutoComplete component}

\description{
Lazy loaded Autocomplete  @param { 	id, 	autoselect, 	cssNamespace, 	defaultValue, 	displayMenu, 	minLength, 	name, 	placeholder, 	onConfirm, 	confirmOnBlur, 	showNoOptionsFound, 	showAllValues, 	required, 	tNoResults, 	tAssistiveHint, 	source, 	templates, 	dropdownArrow: dropdownArrowFactory, 	tStatusQueryTooShort, 	tStatusNoResults, 	tStatusSelectedOption, 	tStatusResults, } [props={}] @return {*}
}

\usage{
'govuk'AutoComplete(id=NULL, ariaLabelledBy=NULL, autoselect=NULL,
confirmOnBlur=NULL, cssNamespace=NULL, displayMenu=NULL,
dropdownArrow=NULL, element=NULL, minLength=NULL, name=NULL,
onConfirm=NULL, placeholder=NULL, required=NULL,
showAllValues=NULL, showNoOptionsFound=NULL, source=NULL,
tAssistiveHint=NULL, tNoResults=NULL, tStatusNoResults=NULL,
tStatusQueryTooShort=NULL, tStatusResults=NULL,
tStatusSelectedOption=NULL, templates=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{ariaLabelledBy}{Character. This is the ID for the label field}

\item{autoselect}{Logical. Should auto select}

\item{confirmOnBlur}{Logical. No Description}

\item{cssNamespace}{Character. cssNamespace}

\item{displayMenu}{Character. No Description}

\item{dropdownArrow}{Logical | numeric | character | named list | unnamed list. React component for dropdown arrow}

\item{element}{Logical | numeric | character | named list | unnamed list. Accessible element}

\item{minLength}{Numeric. No Description}

\item{name}{Character. No Description}

\item{onConfirm}{No Description}

\item{placeholder}{Character. No Description}

\item{required}{Logical. No Description}

\item{showAllValues}{Logical. No Description}

\item{showNoOptionsFound}{Logical. No Description}

\item{source}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tAssistiveHint}{No Description}

\item{tNoResults}{No Description}

\item{tStatusNoResults}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tStatusQueryTooShort}{No Description}

\item{tStatusResults}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tStatusSelectedOption}{No Description}

\item{templates}{Logical | numeric | character | named list | unnamed list. No Description}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

