% Auto-generated: do not edit by hand
\name{'govuk'ComboBox}

\alias{'govuk'ComboBox}

\title{ComboBox component}

\description{

}

\usage{
'govuk'ComboBox(id=NULL, allowsCustomValue=NULL, defaultInputValue=NULL,
defaultItems=NULL, items=NULL, label=NULL, menuTrigger=NULL,
onInputChange=NULL, onSelectionChange=NULL, selectedKey=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{allowsCustomValue}{Logical. }

\item{defaultInputValue}{Character. }

\item{defaultItems}{Logical | numeric | character | named list | unnamed list. }

\item{items}{List | list. }

\item{label}{Character. }

\item{menuTrigger}{Character. }

\item{onInputChange}{}

\item{onSelectionChange}{}

\item{selectedKey}{Character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

