% Auto-generated: do not edit by hand
\name{'govuk'CurrentTime}

\alias{'govuk'CurrentTime}

\title{CurrentTime component}

\description{
Lazy loaded ComponentTemplate  @param { 	id,  value, } [props={}] @return {*}
}

\usage{
'govuk'CurrentTime(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{value}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

