% Auto-generated: do not edit by hand
\name{'govuk'Dropdown}

\alias{'govuk'Dropdown}

\title{Dropdown component}

\description{
@param {{ 	label: string, 	id, 	autoselect, 	cssNamespace, 	value, 	minLength, 	name, 	placeholder, 	onConfirm, 	confirmOnBlur, 	showNoOptionsFound, 	required, 	tNoResults, 	tAssistiveHint, 	source, 	templates, 	dropdownArrow: dropdownArrowFactory, 	tStatusQueryTooShort, 	tStatusNoResults, 	tStatusSelectedOption, 	tStatusResults,  errorMessage, }} [props={}]  @return {*}
}

\usage{
'govuk'Dropdown(id=NULL, autoselect=NULL, confirmOnBlur=NULL,
cssNamespace=NULL, dropdownArrow=NULL, errorMessage=NULL,
label=NULL, minLength=NULL, name=NULL, onConfirm=NULL,
placeholder=NULL, required=NULL, selectElement=NULL,
showNoOptionsFound=NULL, source=NULL, style=NULL,
tAssistiveHint=NULL, tNoResults=NULL, tStatusNoResults=NULL,
tStatusQueryTooShort=NULL, tStatusResults=NULL,
tStatusSelectedOption=NULL, templates=NULL, value=NULL,
wrapperRef=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{autoselect}{Logical. Should auto select}

\item{confirmOnBlur}{Logical. No Description}

\item{cssNamespace}{Character. cssNamespace}

\item{dropdownArrow}{Logical | numeric | character | named list | unnamed list. React component for dropdown arrow}

\item{errorMessage}{Character. Error message to display when invalid input entered in dropdown}

\item{label}{Character. Label text}

\item{minLength}{Numeric. No Description}

\item{name}{Character. No Description}

\item{onConfirm}{No Description}

\item{placeholder}{Character. No Description}

\item{required}{Logical. No Description}

\item{selectElement}{Logical | numeric | character | named list | unnamed list. Accessible element}

\item{showNoOptionsFound}{Logical. No Description}

\item{source}{Logical | numeric | character | named list | unnamed list. No Description}

\item{style}{Logical | numeric | character | named list | unnamed list. Override the css style of the wrapper}

\item{tAssistiveHint}{No Description}

\item{tNoResults}{No Description}

\item{tStatusNoResults}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tStatusQueryTooShort}{No Description}

\item{tStatusResults}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tStatusSelectedOption}{No Description}

\item{templates}{Logical | numeric | character | named list | unnamed list. No Description}

\item{value}{Character. The value displayed in the input.}

\item{wrapperRef}{Logical | numeric | character | named list | unnamed list. wrapperRef}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

