% Auto-generated: do not edit by hand
\name{'govuk'dropdownArrowDown}

\alias{'govuk'dropdownArrowDown}

\title{dropdownArrowDown component}

\description{
efault dropdown arrow
}

\usage{
'govuk'dropdownArrowDown(className=NULL)
}

\arguments{
\item{className}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

