% Auto-generated: do not edit by hand
\name{'govuk'ListBox}

\alias{'govuk'ListBox}

\title{ListBox component}

\description{
Listbox  @param {*} [props={}] @return {*}
}

\usage{
'govuk'ListBox(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

