% Auto-generated: do not edit by hand
\name{'govuk'Popover}

\alias{'govuk'Popover}

\title{Popover component}

\description{
Propover  @export @param {*} props @return {*}
}

\usage{
'govuk'Popover(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

