% Auto-generated: do not edit by hand
\name{'govuk'ExpandableMenuItem}

\alias{'govuk'ExpandableMenuItem}

\title{ExpandableMenuItem component}

\description{
Lazy loaded ExpandableMenuItem
  @param {
 	id,
  title,
  collaExpandableMenuItem
 } [props={}]
 @return {*}
}

\usage{
'govuk'SideNavigation(id=NULL, collapsedByDefault=NULL, title=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components

in callbacks. The ID needs to be unique across all of the

components in an app.}

\item{collapsedByDefault}{Logical. }

\item{title}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

