% Auto-generated: do not edit by hand
\name{'govuk'status}

\alias{'govuk'status}

\title{status component}

\description{
@description Show the status for Autocomplete
}

\usage{
'govuk'status(id=NULL, isInFocus=NULL, length=NULL, minQueryLength=NULL,
queryLength=NULL, selectedOption=NULL,
selectedOptionIndex=NULL, tNoResults=NULL,
tQueryTooShort=NULL, tResults=NULL, tSelectedOption=NULL,
validChoiceMade=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{isInFocus}{Logical. }

\item{length}{Numeric. }

\item{minQueryLength}{Numeric. }

\item{queryLength}{Numeric. }

\item{selectedOption}{Character. }

\item{selectedOptionIndex}{Numeric. }

\item{tNoResults}{}

\item{tQueryTooShort}{}

\item{tResults}{}

\item{tSelectedOption}{}

\item{validChoiceMade}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

