% Auto-generated: do not edit by hand
\name{'govuk'Team2}

\alias{'govuk'Team2}

\title{Team2 component}

\description{
Lazy loaded ComponentTemplate  @param { 	id, } [props={}] @return {*}
}

\usage{
'govuk'Team2(id=NULL, clicks=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{clicks}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

