% Auto-generated: do not edit by hand
\name{govukComponentTemplate}

\alias{govukComponentTemplate}

\title{ComponentTemplate component}

\description{
Lazy loaded ComponentTemplate  @param { 	id, } [props={}] @return {*}
}

\usage{
govukComponentTemplate(id=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

