% Auto-generated: do not edit by hand
\name{govukdropdownArrowDown}

\alias{govukdropdownArrowDown}

\title{dropdownArrowDown component}

\description{
Default dropdown arrow  @param {string} { className }
}

\usage{
govukdropdownArrowDown(className=NULL)
}

\arguments{
\item{className}{Character. Class name to add to SVG}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

