% Auto-generated: do not edit by hand
\name{govukListBox}

\alias{govukListBox}

\title{ListBox component}

\description{
Listbox  @param {*} [props={}] @return {*}
}

\usage{
govukListBox(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

