% Auto-generated: do not edit by hand
\name{govukPopover}

\alias{govukPopover}

\title{Popover component}

\description{
Propover  @export @param {*} props @return {*}
}

\usage{
govukPopover(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

