% Auto-generated: do not edit by hand
\name{govukstatus}

\alias{govukstatus}

\title{status component}

\description{
Display ststus  @export @class Status @extends {Component}
}

\usage{
govukstatus(id=NULL, length=NULL, minQueryLength=NULL, queryLength=NULL,
selectedOption=NULL, selectedOptionIndex=NULL,
tNoResults=NULL, tQueryTooShort=NULL, tResults=NULL,
tSelectedOption=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{length}{Numeric. length}

\item{minQueryLength}{Numeric. min Query Length}

\item{queryLength}{Numeric. Query Length}

\item{selectedOption}{Logical | numeric | character | named list | unnamed list. Selected option}

\item{selectedOptionIndex}{Numeric. Selected Option Index}

\item{tNoResults}{tNoResults}

\item{tQueryTooShort}{tQueryTooShort}

\item{tResults}{tResults}

\item{tSelectedOption}{tSelectedOption}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

