% Auto-generated: do not edit by hand
\name{'govuk'Accordion}

\alias{'govuk'Accordion}

\title{Accordion component}

\description{
Lazy loaded ComponentTemplate  @param { 	id,  accordionHeadings,  children, } [props={}] @return {*}
}

\usage{
'govuk'Accordion(children=NULL, id=NULL, accordionHeadings=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers. Array of accordion children.}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{accordionHeadings}{List of characters. Array of accordion headings.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

