% Auto-generated: do not edit by hand
\name{'govuk'AutoComplete}

\alias{'govuk'AutoComplete}

\title{AutoComplete component}

\description{
Lazy loaded Autocomplete  @param { 	id, 	autoselect, 	cssNamespace, 	defaultValue, 	displayMenu, 	minLength, 	name, 	placeholder, 	onConfirm, 	confirmOnBlur, 	showNoOptionsFound, 	showAllValues, 	required, 	tNoResults, 	tAssistiveHint, 	source, 	templates, 	dropdownArrow: dropdownArrowFactory, 	tStatusQueryTooShort, 	tStatusNoResults, 	tStatusSelectedOption, 	tStatusResults, 	style,  rerender, } [props={}] @return {*}
}

\usage{
'govuk'AutoComplete(id=NULL, alwaysDisplayArrow=NULL, ariaLabelledBy=NULL,
autoselect=NULL, confirmOnBlur=NULL, cssNamespace=NULL,
displayMenu=NULL, dropdownArrow=NULL, minLength=NULL,
name=NULL, onConfirm=NULL, placeholder=NULL, required=NULL,
rerender=NULL, selectElement=NULL, showAllValues=NULL,
showNoOptionsFound=NULL, source=NULL, style=NULL,
tAssistiveHint=NULL, tNoResults=NULL, tStatusNoResults=NULL,
tStatusQueryTooShort=NULL, tStatusResults=NULL,
tStatusSelectedOption=NULL, templates=NULL, value=NULL,
wrapperRef=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{alwaysDisplayArrow}{Logical. alwaysDisplayArrow}

\item{ariaLabelledBy}{Character. This is the ID for the label field}

\item{autoselect}{Logical. Should auto select}

\item{confirmOnBlur}{Logical. No Description}

\item{cssNamespace}{Character. cssNamespace}

\item{displayMenu}{Character. No Description}

\item{dropdownArrow}{Logical | numeric | character | named list | unnamed list. React component for dropdown arrow}

\item{minLength}{Numeric. No Description}

\item{name}{Character. No Description}

\item{onConfirm}{No Description}

\item{placeholder}{Character. No Description}

\item{required}{Logical. No Description}

\item{rerender}{Logical. Force a re render of the component}

\item{selectElement}{Logical | numeric | character | named list | unnamed list. Accessible element}

\item{showAllValues}{Logical. No Description}

\item{showNoOptionsFound}{Logical. No Description}

\item{source}{Logical | numeric | character | named list | unnamed list. No Description}

\item{style}{Logical | numeric | character | named list | unnamed list. Override the css style of the wrapper}

\item{tAssistiveHint}{No Description}

\item{tNoResults}{No Description}

\item{tStatusNoResults}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tStatusQueryTooShort}{No Description}

\item{tStatusResults}{Logical | numeric | character | named list | unnamed list. No Description}

\item{tStatusSelectedOption}{No Description}

\item{templates}{Logical | numeric | character | named list | unnamed list. No Description}

\item{value}{Character. The value displayed in the input.}

\item{wrapperRef}{Logical | numeric | character | named list | unnamed list. wrapperRef}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

