% Auto-generated: do not edit by hand
\name{'govuk'dropdownArrowDown}

\alias{'govuk'dropdownArrowDown}

\title{dropdownArrowDown component}

\description{
Default dropdown arrow  @param {string} { className }
}

\usage{
'govuk'dropdownArrowDown(className=NULL)
}

\arguments{
\item{className}{Character. Class name to add to SVG}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

