% Auto-generated: do not edit by hand
\name{'govuk'ExpandableMenuItem}

\alias{'govuk'ExpandableMenuItem}

\title{ExpandableMenuItem component}

\description{
Lazy loaded ExpandableMenuItem  @param { 	id,  title,  collapsedByDefault,  children, } [props={}] @return {*}
}

\usage{
'govuk'ExpandableMenuItem(children=NULL, id=NULL, collapsedByDefault=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. An array of li HTML elements that will displayed on click}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{collapsedByDefault}{Logical. Default behaviour of whether or not the sub-menu is collapsed on load}

\item{title}{Character. The clickable text to display sub-menu}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

